+++ src/tools/Command/GroupAddTools.ts
@@ -3,20 +3,23 @@ import type { TAhkTokenLine } from '../../globalEnum';
 import type { TScanData } from '../DeepAnalysis/FnVar/def/spiltCommandAll';
 import { spiltCommandAll } from '../DeepAnalysis/FnVar/def/spiltCommandAll';
 
-function getMenuFuncData(lStr: string, col: number): TScanData | null {
+function getGroupAddData_core(lStr: string, col: number, findLabelData: boolean): TScanData | null {
     const strF: string = lStr
         .slice(col)
         .replace(/^\s*GroupAdd\b\s*(?:,\s*)?/iu, 'GroupAdd,')
         .padStart(lStr.length, ' ');
 
     const arr: TScanData[] = spiltCommandAll(strF);
-    const a4: TScanData | undefined = arr.at(4);
-    if (a4 === undefined) return null;
+    const find = findLabelData
+        ? 4
+        : 1;
+    const ax: TScanData | undefined = arr.at(find);
+    if (ax === undefined) return null;
 
     // GroupAdd, GroupName [, WinTitle, WinText, Label, ExcludeTitle, ExcludeText]
     // a0          a1        a2         a3        a4
 
-    const { RawNameNew, lPos } = a4;
+    const { RawNameNew, lPos } = ax;
 
     if (RawNameNew.includes('%')) return null;
 
@@ -32,7 +35,7 @@ function getMenuFuncData(lStr: string, col: number): TScanData | null {
  * ;                                        ^^^^^
  * ```
  */
-export function getGroupAddFunc(AhkTokenLine: TAhkTokenLine): TScanData | null {
+export function getGroupAddData(AhkTokenLine: TAhkTokenLine, findLabelData: boolean): TScanData | null {
     const {
         fistWordUp,
         fistWordUpCol,
@@ -41,8 +44,8 @@ export function getGroupAddFunc(AhkTokenLine: TAhkTokenLine): TScanData | null {
         SecondWordUpCol,
     } = AhkTokenLine;
 
-    if (fistWordUp === 'GROUPADD') return getMenuFuncData(lStr, fistWordUpCol);
-    if (SecondWordUp === 'GROUPADD') return getMenuFuncData(lStr, SecondWordUpCol);
+    if (fistWordUp === 'GROUPADD') return getGroupAddData_core(lStr, fistWordUpCol, findLabelData);
+    if (SecondWordUp === 'GROUPADD') return getGroupAddData_core(lStr, SecondWordUpCol, findLabelData);
 
     return null;
 }